views controller structure changed
